pyproject.toml
[tool.poetry]
name = "gptdocsbot"
version = "0.1.0"
description = "Talk to a GPT4 bot using langchain that can query code documentation"
authors = ["Marcus Weinberger <marcusjw@duck.com>"]

[tool.poetry.dependencies]
python = "^3.10"
gpt-index = "^0.5.6"
langchain = "^0.0.130"
openai = "^0.27.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
